requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.1
billiard==3.6.4.0
celery==5.2.7
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Command==0.1.0
cron-descriptor==1.2.35
cryptography==36.0.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.4
Django==4.2
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-debug-toolbar==3.2.4
django-debug-toolbar-template-profiler==2.0.2
django-extensions==3.1.5
django-timezone-field==5.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
filelock==3.4.0
geographiclib==2.0
geopy==2.3.0
h11==0.14.0
idna==3.3
ipython==7.29.0
jedi==0.18.1
kombu==5.2.4
matplotlib-inline==0.1.3
numpy==1.24.2
oauthlib==3.1.1
openpyxl==3.1.2
outcome==1.2.0
pandas==1.5.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
prompt-toolkit==3.0.22
psycopg2==2.9.5
pycparser==2.21
Pygments==2.10.0
PyJWT==2.3.0
PySocks==1.7.1
python-crontab==2.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
redis==4.5.4
requests==2.26.0
requests-oauthlib==1.3.0
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sortedcontainers==2.4.0
sqlparse==0.4.2
traitlets==5.1.1
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
urllib3==1.26.7
vine==5.0.0
virtualenv==20.10.0
wcwidth==0.2.5
wrapt==1.13.3
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
